certifier/pyproject.toml
[tool.poetry]
name = "certifier"
version = "0.2.0"
description = "self-signed certificates generator 🔒"
readme = "README.md"
authors = ["radio-aktywne <it@radioaktywne.pl>"]
classifiers = ["License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"]
repository = "https://github.com/radio-aktywne/certifier"
documentation = "https://radio-aktywne.github.io/certifier"

[tool.poetry.dependencies]
python = "^3.9"
typer = { version = "^0.4", extras = ["all"] }
trustme = "^0.9"
pydantic = "^1.10"
pyhumps = "^3.7"
PyYAML = "^6.0"
deepmerge = "^1.0"
httpx = "^0.23"

# dev

# test
pytest = { version = "^7.0", optional = true }

[tool.poetry.extras]
# need to do it that way until poetry supports dependency groups: https://github.com/python-poetry/poetry/issues/1644
dev = ["pytest"]
test = ["pytest"]
[tool.poetry.scripts]
# cli entry point
certifier = "certifier.__main__:cli"

[build-system]
# this should be there, see https://python-poetry.org/docs/pyproject/#poetry-and-pep-517
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
